Replacing Predictor with BasePredictor
More flexible for separate stage1 and stage2 predictors
